+++ src/main/java/reposense/model/ViewCliArguments.java
@@ -2,8 +2,17 @@ package reposense.model;
 
 import java.nio.file.Path;
 
+/**
+ * Represents command line arguments user supplied when running the program with mandatory field -view.
+ */
 public class ViewCliArguments extends CliArguments {
+    private Path reportDirectoryPath;
+
     public ViewCliArguments(Path reportDirectoryPath) {
         this.reportDirectoryPath = reportDirectoryPath;
+    }
+
+    public Path getReportDirectoryPath() {
+        return reportDirectoryPath;
     }
 }
